// React/app/GerAdv_TS/Oponentes/Interfaces/interface.Oponentes.ts
// 100% auto-generated by Source Genesys WebApi Front & Back Creator
                     // copyright © 2000-2025 Menphis - Sistemas Inteligentes
                     // This file is part of the Source Genesys project                     
                    'use client';

export interface IOponentes {
// 202501251
    id: number;
 
	cidade: number,
	empfuncao: number,
	ctpsnumero: string,
	site: string,
	ctpsserie: string,
	nome: string,
	adv: number,
	empcliente: number,
	idrep: number,
	pis: string,
	contato: string,
	cnpj: string,
	rg: string,
	juridica: boolean,
	tipo: boolean,
	sexo: boolean,
	cpf: string,
	endereco: string,
	fone: string,
	fax: string,
	bairro: string,
	cep: string,
	inscest: string,
	observacao: string,
	email: string,
	class: string,
	top: boolean,
}

export interface IOponentesFormProps {
  id: number;
  onClose: () => void;
  onSuccess: (registro?: any) => void;
  onError: () => void;
}

export interface IOponentesIncProps {
    id: number;
    onClose: () => void;
    onSuccess: (registro?: any) => void;
    onError: () => void;
}